Move common test to parent class and reuseForks=false to test with different system properties
