+++ jbpm-seldon-prediction-service/pom.xml
@@ -62,6 +62,12 @@
                         </execution>
                     </executions>
                 </plugin>
+                <plugin>
+                     <artifactId>maven-surefire-plugin</artifactId>
+                     <configuration>
+                        <reuseForks>false</reuseForks>
+                     </configuration>
+                 </plugin>
             </plugins>
         </pluginManagement>
     </build>
@@ -219,6 +225,5 @@
             <artifactId>wiremock-jre8-standalone</artifactId>
             <scope>test</scope>
         </dependency>
-
     </dependencies>
 </project>
