+++ jbpm-seldon-prediction-service/src/test/java/org/jbpm/prediction/service/seldon/NDArrayResponseTest.java
@@ -18,28 +18,11 @@ package org.jbpm.prediction.service.seldon;
 
 import org.junit.Test;
 
-import java.util.Map;
-
-import static com.github.tomakehurst.wiremock.client.WireMock.*;
-import static org.junit.Assert.assertEquals;
-
-public class NDArrayResponseTest extends AbstractSeldonTestSuite {
+public class NDArrayResponseTest extends SeldonTestSuite {
 
     @Test
     public void testNDArrayResponse() {
-        final String endpoint = System.getProperty("org.jbpm.task.prediction.service.seldon.endpoint");
-        stubFor(post(urlEqualTo("/" + endpoint))
-                .withHeader("Accept", equalTo("application/json"))
-                .willReturn(aResponse()
-                        .withStatus(200)
-                        .withHeader("Content-Type", "application/json")
-                        .withBody(TestUtils.readJSONAsString("responses/ndarray.json"))));
-
-        startAndReturnTaskOutputData("test item", "john", 5, false);
-        Map<String, Object> outputs = startAndReturnTaskOutputData("test item", "john", 5, true);
-
-        final double confidence = (double) outputs.get("confidence");
-        assertEquals(0.71, confidence, 1e-10);
+        super.testNDArrayResponse();
     }
 
 }
